Chore: Move files to have a cleaner file structure
