+++ admin/routes.json
@@ -1,14 +1,14 @@
 {
-  "admin.app.returns": {
-    "component": "ReturnsAdmin",
+  "admin.app.return-list": {
+    "component": "AdminReturnList",
     "path": "/admin/app/returns/requests/"
   },
-  "admin.app.returnsSettings": {
-    "component": "RMASettings",
+  "admin.app.return-settings": {
+    "component": "AdminReturnSettings",
     "path": "/admin/app/returns/settings/"
   },
-  "admin.app.returnDetails": {
-    "component": "ReturnDetails",
+  "admin.app.return-details": {
+    "component": "AdminReturnDetails",
     "path": "/admin/app/returns/:id/details/"
   }
 }
