+++ react/AdminReturnSettings.tsx
