+++ react/StoreMyReturnsLink.json
@@ -1,5 +1,5 @@
 [
   {
-    "id": "returns.link"
+    "id": "store/return-app.link"
   }
 ]
