+++ react/store/ReturnsDetails.tsx
@@ -32,7 +32,6 @@ const messages = defineMessages({
 class ReturnsDetails extends Component<any, any> {
   static propTypes = {
     headerConfig: PropTypes.object,
-    fetchApi: PropTypes.func,
   }
 
   constructor(props: any) {
