+++ react/store/StoreMyReturnsDetails.tsx
@@ -1,11 +1,9 @@
 import type { FC } from 'react'
 import React from 'react'
-import { Route } from 'vtex.my-account-commons/Router'
 import { FormattedMessage } from 'react-intl'
 import { useRuntime } from 'vtex.render-runtime'
 
-import ReturnsDetails from './store/ReturnsDetails'
-import useAxiosInstance from './hooks/useAxiosModule'
+import ReturnsDetails from './ReturnsDetails'
 
 const headerConfig = {
   namespace: 'vtex-account__returns-list',
@@ -16,29 +14,15 @@ const headerConfig = {
   },
 }
 
-const StoreMyReturnsDetailsWrapper: FC = (props: any) => {
-  const fetchApi = useAxiosInstance()
+export const StoreMyReturnsDetailsWrapper: FC = (props: any) => {
   const { production, binding } = useRuntime()
 
   return (
     <ReturnsDetails
       {...props}
       headerConfig={headerConfig}
-      fetchApi={fetchApi}
       production={production}
       binding={binding}
     />
   )
 }
-
-const StoreMyReturnsDetails = () => {
-  return (
-    <Route
-      exact
-      path="/my-returns/details/:id"
-      component={StoreMyReturnsDetailsWrapper}
-    />
-  )
-}
-
-export default StoreMyReturnsDetails
