+++ react/store/StoreMyReturnsPageAdd.tsx
@@ -1,14 +1,12 @@
 import type { FC } from 'react'
-import React, { Fragment } from 'react'
-import { Route } from 'vtex.my-account-commons/Router'
+import React from 'react'
 import { FormattedMessage, useIntl } from 'react-intl'
 import { useRuntime } from 'vtex.render-runtime'
 import { useMutation } from 'react-apollo'
 import { useCssHandles } from 'vtex.css-handles'
 
-import useAxiosInstance from './hooks/useAxiosModule'
-import ReturnsPageAdd from './store/MyReturnsPageAdd'
-import CREATE_RETURN_REQUEST from './store/graphql/createReturnRequest.gql'
+import ReturnsPageAdd from './MyReturnsPageAdd'
+import CREATE_RETURN_REQUEST from './graphql/createReturnRequest.gql'
 
 const headerConfig = {
   namespace: 'vtex-account__returns-list',
@@ -21,8 +19,7 @@ const headerConfig = {
 
 const CSS_HANDLES = ['errorContainer']
 
-const StoreMyReturnsPageAddWrapper: FC = (props: any) => {
-  const axios = useAxiosInstance()
+export const StoreMyReturnsPageAddWrapper: FC = (props: any) => {
   const { production, binding, rootPath } = useRuntime()
   const intl = useIntl()
   const [
@@ -53,7 +50,6 @@ const StoreMyReturnsPageAddWrapper: FC = (props: any) => {
     <ReturnsPageAdd
       {...props}
       headerConfig={headerConfig}
-      fetchApi={axios}
       production={production}
       binding={binding}
       rootPath={rootPath}
@@ -66,15 +62,3 @@ const StoreMyReturnsPageAddWrapper: FC = (props: any) => {
     />
   )
 }
-
-const StoreMyReturnsPageAdd = () => (
-  <Fragment>
-    <Route
-      exact
-      path="/my-returns/add"
-      component={StoreMyReturnsPageAddWrapper}
-    />
-  </Fragment>
-)
-
-export default StoreMyReturnsPageAdd
