+++ store/interfaces.json
@@ -2,13 +2,7 @@
   "my-account-link.my-returns-link": {
     "component": "StoreMyReturnsLink"
   },
-  "my-account-link.my-returns-page": {
-    "component": "StoreMyReturnsPage"
-  },
-  "my-account-link.my-returns-page-add": {
-    "component": "StoreMyReturnsPageAdd"
-  },
-  "my-account-link.my-returns-page-details": {
-    "component": "StoreMyReturnsDetails"
+  "my-account-link.my-returns-app": {
+    "component": "StoreReturnAppRouter"
   }
 }
